Attempt to increase logic thread performance / FPS
Testing some optimisations to the position integration with respect to
time for planets (the physics calculations which occur in the logic
thread). A greater FPS / smaller timestep leads to a more accurate
simulation.

+++ src/physics/Constants.java
@@ -2,6 +2,6 @@ package physics;
 
 public class Constants {
 	
-	public static final double G = 6.673E-2;
+	public static final float G = 6.673E-2f;
 	
 }
